Factor out common executor utils
